Adjust metrics to avoid Stopwatch allocation using .NET 7 pattern
